Funciona (sin SnakeFront)

+++ Position.h
@@ -2,8 +2,8 @@
 #define _Position_H_
 
 typedef struct {
-  uint8_t x; //Horizontal 0=izquierda
-  uint8_t y; //Vertical 0=abajo
+  int x; //Horizontal 0=izquierda
+  int y; //Vertical 0=abajo
 } Position;
 
 #endif
